@@ -1,5 +1,7 @@
 class CiMailer < ActionMailer::Base
-  default from: "zemagno@gmail.com"
+    #AbstractController
+    append_view_path SqlTemplate::Resolver.new
+    default from: "zemagno@gmail.com"
 
   # Subject can be set in your I18n file at config/locales/en.yml
   # with the following lookup:
@@ -15,8 +17,9 @@ class CiMailer < ActionMailer::Base
     @ci = obj
     mail to: destinatario
   end
+
   def problema_servidor(obj,subject,destinatario)
-    @ci = obk
+    @ci = obj
     mail to: destinatario
   end
 
